Truncate subcommand description after first paragraph in usage
Previously we showed the subcommand description in its entirety in the
list of subcommands of a branch. That was way too much.

Put some shine on the examples in light of the truncation. By convention
the first paragraph is a single imperative statement e.g. "Display
detailed information about a user"