Implement new method ICli#runLine
This new method parses the provided line and splits it into args which
it then passes into ICli#run. The line-splitting logic is promoted from
the website module where it was a part of the CommandLine class. The
runLine method seems like it will be the most convenient way to run a
browser console CLI.

Flesh out main package readme as face of the documentation. Add
screenshots of Node.js and browser console CLIs.

Commit the readme examples to the examples/website packages.

Publish 0.5.0-22 and 2020.7.3